app/src/main/java/com/marvim/wishlist/application/usecases/AddProductToWishlistUseCaseImpl.java
package com.marvim.wishlist.application.usecases;

import com.marvim.wishlist.domain.entity.Product;
import com.marvim.wishlist.domain.ports.input.AddProductToWishlistUseCase;
import com.marvim.wishlist.domain.ports.output.WishlistRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class AddProductToWishlistUseCaseImpl implements AddProductToWishlistUseCase {

    private final WishlistRepository wishlistRepository ;

    @Override
    public void execute(String clientId, Product product) {
        wishlistRepository.add(product);
    }
}
